1.A lot of small adjustments 2.Started to work on new before stage action

+++ src/gameplay_parts/StatsTracker.cs
@@ -19,7 +19,7 @@ namespace SaYSpin.src.gameplay_parts
             double initShopPriceCoefficient,
             int initTileItemsInShopCount,
             int initRelicsInShopCount,
-            int initCoinsNeededToCompleteStage
+            double initCoinsNeededToCompleteStage
             )
         {
             initialValues = new Dictionary<GameStat, double>
